requirements.txt
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
mysqlclient==2.2.7
pytest==8.3.4 
pytest-django==4.10.0